add error handling

+++ rails/app/models/article.rb
@@ -33,7 +33,9 @@ class Article < ApplicationRecord
       self.message = new_message
       # Message.create(content: self.content, title: self.title, link: "fsf://fsf/news/article/" + self.id.to_s, article_id: self.id)  
     else 
-      self.message.destroy
+      if self.message
+        self.message.destroy
+      else
     end
   end
